fix(LINGUIST-410): Refresh user gems after item purchase
